Mostly working Electron flicker-free scroll
Lightly tested (see TODOs, though) and seems to work on Electron and
B+64K, so I probably haven't broken anything too badly.

We are *really* getting tight for space and the code organisation is
already suffering a bit as a result. I don't really want to give up any
of the loop unrolling. We could maybe squeeze another 6-ish bytes out on
the Electron (which is what - modulo debug colour bar desires - is
really tight for space right now) if we factored out more of the
variable code into one contiguous chunk (using macros to de-duplicate
the actually common bits at source level), accepted the poor report from
acme as a result and the slightly larger on-disc executable. The 6 bytes
would come from saving jmps. Speculating. This may be OK as is - but if
I want/need to do cursor-related stuff for the Electron in this code
that is likely to get painful in terms of code size.

It is nice to see the Electron (in MAME, FWIWI) doing a flicker free
status line in the benchmark in mode 6. (Haven't tried it in mode 3
yet.)