Probably "proper" fix for ugly cursor position
See the removed comments for a description of the problem and the old
fix. I think this new fix is better as it addresses the underlying issue
- the cursor could be visible (if we're printing user input, rather than
the game's own messages) while it got moved into "incorrect-looking"
locations, and the right fix is to make sure the cursor is hidden at
these times. I think the rule is "if you're setting
s_cursors_inconsistent to non-0, the cursor should not be visible", and
I think we now respect that rule.

I have lightly tested this so fingers crossed it does work OK.