Default to OSRDCH on Z1-3, rename --osrdch to --force-osrdch
Lightly tested and seems to work. Renaming the option will make it a bit
more obvious to users who are used to doing --osrdch to work around
emulator problems with data-in-stack that something has changed, rather
than the option being accepted but not having the effects they expect on
data-in-stack (since that's now been pulled out into a distinct option).