Fix bug setting max_lines for Z1/2
Found during merge and subsequent review, this hasn't been reported and
I don't have an example of it going wrong. It's likely I introduced this
bug in the Acorn port when refactoring the upstream code slightly, but I
haven't tried to dig through the upstream history to see if it ever had
the same error.