Finish merge of upstream commit from 608ac1efe
This relates to handling a lower window of just one line high.

I had previously picked up the change of sbc #2 to sbc #1 just above but
not this. I only half understand this change TBH but there is nothing
Acorn-specific going on here so I will trust upstream. (It was almost
certainly broken with just half the change, although of course I am
still in the merge fix up stage so this never got released.)