Delete outdated TODO
I presumably did test this before the fast hardware scrolling release,
which was a while ago now. But to be safe I just ran the start of the
benchmark in mode 6 with forced software and forced slow hardware
scrolling and it looks fine. (Fast hardware scrolling is the default so
naturally gets tested all the time.)