Move .XINDEX into zero page and rename it
I am 95% confident this is OK and I have lightly tested it and it seems
to work. It feels as though when we've just finished reading input from
the user there shouldn't be much risk of some other code having stored
valuable data in zp_temp - it isn't as if this code is called at
arbitrary points in the middle of complex operations.