Tweak generated files output from make-acorn.py
This should be a bit more generic and machine parseable if/when
make-acorn.py can generate multiple output files, as we can output them
on a single line space-delimited. (The most likely - albeit not that
likely - case I can think of for this is adding support for games on
100K DFS discs where disc swapping is required.)

Thanks to Fredrik for raising this.