Implement make-acorn.py quality of life improvements
Auto-delete temporary files in a hopefully safe way unless explicitly
told to keep them.

Allow make-acorn.py to be run when it isn't in the current directory.

Thanks to Fredrik for bringing these up.

This also tweaks CACHEDIR.TAG handling so we don't recreate the file if
it exists. This probably doesn't matter, but may avoid introducing
gratuitous noise into backups or similar.