Fix extra_main_ram calculation and make sure bigdyn header in main RAM
This is a work in progress. It seems to be working based on my testing
but I need to review it fresh, and as per SFTODONOW I also need to see
if I am incorrectly reporting the extra RAM needed for other memory
models because I am not taking account of natural alignment there.