Remove TODO
I'm fairly sure this is fine. As per deleted comment, by the time we
print anything after the restart re-load, the code has relocated itself
down below screen memory. And the cursor will be off (because it always
is when we're not waiting for input) so the Electron's software cursor
can't cause problems.