Fix recent ReCo6502 turbo test fix
Running TURBO implicitly sets &ED to zero but it doesn't clear &7000, so
if TURBO fails to execute (because we're not on a turbo co-pro) &7000
would have an arbitrary value. We need to explicitly zero it first.

Really it might be easier just to make TURBO run at ~&7000 but let's go
with this for now.