Fix dynmem adjust shrinking nonstored_pages below ACORN_INITIAL_NONSTORED_PAGES
A corner case found with Fredrik's new "catch" test (again, not actually
anything to do with throw/catch, just shown up by its small size). This
feels like a horrible hack but it does seem to work. It *might* be
cleaner to just force a minimum of two pages of VM on the game -
virtually no real game will trigger this, and the extra load time is
negligible for a tiny one - to avoid this and the previous corner case,
but that feels a bit like giving in.