Possibly mostly working global var optimisation
This is turning into a bit of a build-time nightmare of special cases
and I need to run through builds with and without this forced off and
with different dynamic memory models on differently sized target
systems, but I think this *may* be correct. Still some TODOs to tidy up
even if it is.

The tweak to bbc_max_start_addr and the assert <= $7c00 instead of $8000
is because the recent changes meant *LOADing the OZMOOB binary in mode 7
would corrupt screen RAM, which is harmless but ugly. I assume the
underlying cause of this is the increased binary size, even though the
new code is just discardable init code so it isn't really a concern.