Moved saved_* into zp where possible
I did a crude analysis on the acme listing file (ignoring the discardable init code
after program_end) and identified these as good candidates for zero page
on grounds of code size - it looks like this also gives a modest code
boost as well.

I just re-discovered that my memory allocation code is smart enough to
go back and "fill in" the gaps if (say) a 2 byte allocation won't fit in
zp but a later 1 byte allocation will. Go me!