Improve detection of Watford DFS *FX111 use
This is a bigger diff than it should be, because the shadow driver
install code is extremely tight for space and I had to squash it further
to make things fit, which required shuffling some otherwise unmodified
code around.

Lightly tested (not on the relevant emulated hardware yet though!) and
seems to work.