Fix corner case bugs in memory checks
As shown up by Varicella. This stuff does my head in and definitely
needs reviewing fresh later on and ideally some testing with various
games and RAM configurations.

It *might* have been possible to avoid passing data_start through but
I'm not sure, and having this available *does* feel like a fairly
obviously correct way to verify the header is in main RAM for big dynmem
builds without getting unnecessarily clever (as I did before, and it
turned out I was wrong).

data_start == story_start in the cases we care about, but data_start
feels more general and since we have to pass this through for all models
even if we're only using it for big dynmem at the moment, this slightly
eases the path to redoing the memory free calculations in future on (for
example) the basis of the build system telling the loader data_start and
the game's (rounded up to a multiple of 512 byte) dynmem size and
letting it use those values to check. I may be overlooking some
complexities but it feels like that would actually be way simpler.