Don't accidentally force no slow HW scroll on some builds
I was taking a reference to ozmoo_base_args when I meant to copy it, so
the "BBC B no shadow" removal of ACORN_HW_SCROLL_SLOW propagated to
other builds.

I found this while testing via a hacked "screen only shadow driver for
Master" case that it is OK to re-use the fast hardware scroll code
location for slow hardware scrolling when this is done "properly" with
the real fast scroll driver.

I am not mentioning this in the changelog as the "don't include slow HW
scroll support on BBC B no shadow" change only came in in this release.