Move SHADDRV to &2C00
Lightly tested and seems to work. In MAME with Aries B20 and 16K
sideways RAM, the benchmark finishes on $e454 in mode 7 and $1f3a0 in
mode 3, so this suggests spare shadow RAM is being used to cache data
correctly. The "Aries (screen only)" problem when rebooting with a soft
BREAK also seems to be fixed by the addition of the extra "lda #111".