Remove TODO
A quick test (you have to turn DEBUG_BIG_DYNMEM off to build it like
this anyway) suggests Varicella does run on a Master 128 with PAGE=&E00.
I haven't faffed to test with PAGE=&F00 but I'm fairly sure the original
problem was just down to the buggy and over-complex memory calculation
logic which I reworked and this TODO comment was just lingering after I
sorted that out.